clean up tests for asm
